add test for node classes
