+++ src/Node/Node.ts
@@ -1,0 +1,7 @@
+export class SinglyLinkedNode<T> {
+  public next: SinglyLinkedNode<T> | null = null;
+  constructor(public data: T) {}
+}
+export class DoublyLinkedNode<T> extends SinglyLinkedNode<T> {
+  public prev: DoublyLinkedNode<T> | null = null;
+}
