+++ src/Node/index.ts
@@ -1,7 +1,1 @@
-export class SinglyLinkedNode<T> {
-  public next: SinglyLinkedNode<T> | null = null;
-  constructor(public data: T) {}
-}
-export class DoublyLinkedNode<T> extends SinglyLinkedNode<T> {
-  public prev: DoublyLinkedNode<T> | null = null;
-}
+export { SinglyLinkedNode, DoublyLinkedNode } from "./Node";
